Drastic change in MongoCollection
We do not build anymore the `self.key_to_mongo_id` dict mapping `_id`
and `nice_key`. This lead to a very significant speedup becase it was
taking ~1.2 s to build this dictionary. Instead of doing this we are
using the `nice_key`to find items. The `__getitem__` method has been
updated accordingly.